+++ skore/tests/unit/cli/test_cli.py
@@ -32,7 +32,7 @@ def mock_server_manager(monkeypatch):
             self.start_params = {"port": port, "open_browser": open_browser}
 
     mock_manager = MockServerManager()
-    monkeypatch.setattr("skore.project._manage.ServerManager", mock_manager)
+    monkeypatch.setattr("skore.project._launch.ServerManager", mock_manager)
     return mock_manager
