+++ skore/tests/unit/project/test_load.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 from skore.project import Project
-from skore.project._manage import _load
+from skore.project._load import load
 
 
 @pytest.fixture
@@ -18,15 +18,15 @@ def tmp_project_path(tmp_path):
 
 def test_load_no_project():
     with pytest.raises(FileNotFoundError):
-        _load("/empty")
+        load("/empty")
 
 
 def test_load_absolute_path(tmp_project_path):
-    p = _load(tmp_project_path)
+    p = load(tmp_project_path)
     assert isinstance(p, Project)
 
 
 def test_load_relative_path(tmp_project_path):
     os.chdir(tmp_project_path.parent)
-    p = _load(tmp_project_path.name)
+    p = load(tmp_project_path.name)
     assert isinstance(p, Project)
